Move text_from_HTML out of metaphor.common
