+++ tests/static_web/test_extractor.py
@@ -59,37 +59,6 @@ def test_get_subpages_from_HTML(static_web_extractor):
     assert "https://example.com/test" in result
 
 
-# Test for extracting visible text from HTML, with filtering
-def test_get_text_from_HTML_with_filtering(static_web_extractor, test_root_dir: str):
-    html_content = load_text(
-        f"{test_root_dir}/static_web/sample_pages/titles_text.html"
-    )
-    text = static_web_extractor._get_text_from_HTML(html_content)
-    assert "Visible paragraph 1." in text
-    assert "Visible paragraph 2." in text
-    assert "Test Title" not in text
-    assert "Some style" not in text
-    assert "Some script" not in text
-    assert "Some meta" not in text
-    assert "Commented text" not in text
-    assert "Script text" not in text
-    assert "Style text" not in text
-
-
-# Test for extracting title from HTML
-def test_get_title_from_HTML_success(static_web_extractor):
-    html_content = "<html><head><title>Test Title</title></head></html>"
-    title = static_web_extractor._get_title_from_HTML(html_content)
-    assert title == "Test Title"
-
-
-# Test for extracting empty title
-def test_get_title_from_HTML_failure(static_web_extractor):
-    html_content = "<html><head></head><body><h1>Hello World!</h1></body></html>"
-    title = static_web_extractor._get_title_from_HTML(html_content)
-    assert title == ""
-
-
 # Test for making a document
 def test_make_document(static_web_extractor):
     doc = static_web_extractor._make_document(
